Add separate class for C# `ILocalSymbol` tests
