+++ src/Compilers/CSharp/Test/Symbol/Compilation/SemanticModelGetSemanticInfoTests.cs
@@ -5555,124 +5555,6 @@ class C
             Assert.True(semanticInfo.Symbol.GetSymbol().IsFromCompilation(compilation));
         }
 
-        [WorkItem(59709, "https://github.com/dotnet/roslyn/issues/59709")]
-        [Fact]
-        public void LocalForEach()
-        {
-            string sourceCode = @"
-class C
-{
-    void M()
-    {
-        var a = new int[] { 1, 2, 3 };
-        foreach (var x in a)
-        {
-            var y = /*<bind>*/x/*</bind>*/;
-        }
-    }
-}
-";
-            var compilation = CreateCompilation(sourceCode);
-            var semanticInfo = GetSemanticInfoForTest<IdentifierNameSyntax>(compilation);
-
-            var symbol = (ILocalSymbol)semanticInfo.Symbol;
-            Assert.True(symbol.IsForEach);
-        }
-
-        [WorkItem(59709, "https://github.com/dotnet/roslyn/issues/59709")]
-        [Fact]
-        public void LocalForEachAwait()
-        {
-            string sourceCode = @"
-class C
-{
-    void M()
-    {
-        var a = new int[] { 1, 2, 3 };
-        await foreach (var x in a)
-        {
-            var y = /*<bind>*/x/*</bind>*/;
-        }
-    }
-}
-";
-            var compilation = CreateCompilation(sourceCode);
-            var semanticInfo = GetSemanticInfoForTest<IdentifierNameSyntax>(compilation);
-
-            var symbol = (ILocalSymbol)semanticInfo.Symbol;
-            Assert.True(symbol.IsForEach);
-        }
-
-        [WorkItem(59709, "https://github.com/dotnet/roslyn/issues/59709")]
-        [Fact]
-        public void LocalUsing()
-        {
-            string sourceCode = @"
-using System.IO;
-class C
-{
-    void M()
-    {
-        using (var x = new StreamReader(""""))
-        {
-            /*<bind>*/x/*</bind>*/.Read();
-        }
-    }
-}
-";
-            var compilation = CreateCompilation(sourceCode);
-            var semanticInfo = GetSemanticInfoForTest<IdentifierNameSyntax>(compilation);
-
-            var symbol = (ILocalSymbol)semanticInfo.Symbol;
-            Assert.True(symbol.IsUsing);
-        }
-
-        [WorkItem(59709, "https://github.com/dotnet/roslyn/issues/59709")]
-        [Fact]
-        public void LocalUsingAwait()
-        {
-            string sourceCode = @"
-using System.IO;
-class C
-{
-    void M()
-    {
-        await using (var x = new StreamReader(""""))
-        {
-            /*<bind>*/x/*</bind>*/.Read();
-        }
-    }
-}
-";
-            var compilation = CreateCompilation(sourceCode);
-            var semanticInfo = GetSemanticInfoForTest<IdentifierNameSyntax>(compilation);
-
-            var symbol = (ILocalSymbol)semanticInfo.Symbol;
-            Assert.True(symbol.IsUsing);
-        }
-
-        [WorkItem(59709, "https://github.com/dotnet/roslyn/issues/59709")]
-        [Fact]
-        public void LocalUsingDeclaration()
-        {
-            string sourceCode = @"
-using System.IO;
-class C
-{
-    void M()
-    {
-        using var x = new StreamReader("""");
-        /*<bind>*/x/*</bind>*/.Read();
-    }
-}
-";
-            var compilation = CreateCompilation(sourceCode);
-            var semanticInfo = GetSemanticInfoForTest<IdentifierNameSyntax>(compilation);
-
-            var symbol = (ILocalSymbol)semanticInfo.Symbol;
-            Assert.True(symbol.IsUsing);
-        }
-
         [WorkItem(540541, "http://vstfdevdiv:8080/DevDiv2/DevDiv/_workitems/edit/540541")]
         [Fact]
         public void InEnumElementInitializer()
